Fix entry delete for TodoCard
- Remove renderTemplate() from TodoCard.tsx and other small changes
- Add TodoEntry to replace renderTemplate()
  - update AppFragTodo and App to accomodate the above
  - add Local.TodoEntryProps to index.d.ts
