+++ src/react/components/app-specific/ToDoCard/AppFragTodo.tsx
@@ -1,21 +1,10 @@
 import React, { Component, ReactElement, MouseEvent } from 'react';
 
-import { TodoCard, renderTemplate } from './TodoCard';
-
-import { Button } from '../../base/Button/Button';
-import { Paragraph } from '../../base/Paragraph/Paragraph';
-
-import { List } from '../../presentational/List/List';
-import { ListItem } from '../../presentational/List/ListItem/ListItem';
-
-const sayHi = (event: React.MouseEvent<HTMLInputElement>) => {
-  console.log('hi');
-};
+import { TodoCard } from './TodoCard';
+import { TodoEntry } from './TodoEntry';
 
 const initialState: Local.BasicObj = {
-  listEle: [],
-  ['task']     : '',
-  ['priority'] : ''
+  listEle: []
 }
 
 type State = Readonly<typeof initialState>
@@ -23,12 +12,10 @@ type State = Readonly<typeof initialState>
 export class AppFragTodo extends Component<any, State> {
   readonly state: State = initialState;
 
-  handleSubmit = (state: any): any => {
-    const { listEle } = this.state;
-    const arr = renderTemplate(state, listEle.length, this.handleEntryDelete);
+  handleSubmit = (childState: any): any => {
     this.setState((prevState: State) => {
       const { listEle } = prevState;
-      listEle.push(arr);
+      listEle.push(childState);
       return (
         {
           'listEle': listEle,
@@ -41,23 +28,36 @@ export class AppFragTodo extends Component<any, State> {
     const { listEle } = this.state;
     if (event.target instanceof HTMLElement) {
       const { target } = event.target.dataset;
-      const index = parseInt(target);
       this.setState((prevState: State): State => {
         const { listEle: prevList } = prevState;
-        prevList.splice(index, 1);
+        prevList.splice(target, 1);
         return {};
       })
     }
   }
 
+  renderTemplate = (val: Local.BasicObj, index: number): ReactElement<any> => {
+    const { task, priority } = val;
+    return (
+      <TodoEntry
+        key={index}
+        index={`${index}`}
+        task={task}
+        priority={priority}
+        handleDeleteFn={this.handleEntryDelete}
+      />
+    )
+  }
+
   render = (): ReactElement<any, any> => {
-    const { listEle } = this.state
+    const { listEle } = this.state;
+    const toRender = listEle.length > 0 ? listEle.map(this.renderTemplate) : null;
     return (
       <div>
         <TodoCard
           handleSubmitFn={ this.handleSubmit }
         >
-          { listEle ? listEle : null }
+          { toRender }
         </TodoCard>
       </div>
     )
